Add files via upload
In this Dice Roll Game , two dices are present with respective to 2 players. The given dices is rolled by player . The player who has larger number on his dice between 1 to 6 will be the winner of that turn in game . If both have same number on dice, then the game will be tie up. We have to start from begin. Let's Roll a Dice to Play a Game!